backend/requirements.txt
alembic==1.17.1
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2025.2
qrcode==8.2
requests==2.32.5
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
pytest==7.4.3
pytest-flask==1.2.0
